# first_project/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My First API",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/products/manyProducts/{number}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "get all products",
                "operationId": "537cb0a25e7f64f183e6d438481150da",
                "parameters": [
                    {
                        "name": "cursor",
                        "in": "query",
                        "description": "Cursor for pagination",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "number",
                        "in": "path",
                        "description": "Product number",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful get all products"
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            }
        },
        "/products/oneProduct/{product_id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "get one product",
                "operationId": "261967d6ce1e2750c39fc39cb97b2359",
                "parameters": [
                    {
                        "name": "product_id",
                        "in": "path",
                        "description": "Product id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful get product"
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            }
        },
        "/products/filterBy/category/{category_id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "get products by category id",
                "operationId": "d0ecbefbc2b0fa8d07e916da28ba4861",
                "parameters": [
                    {
                        "name": "category_id",
                        "in": "path",
                        "description": "Category id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful filter products by category"
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            }
        },
        "/products/filterBy/expirationDate/{expiration_date}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "get products by expiration date",
                "operationId": "4a3b7fdf9213454c90ab3f7109fe3c60",
                "parameters": [
                    {
                        "name": "expiration_date",
                        "in": "path",
                        "description": "Category id",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful filter products by expiration date"
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            }
        },
        "/products/filterBy/name/{product_name}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "get products by name",
                "operationId": "aacae4286391f17ad64bcc8793203261",
                "parameters": [
                    {
                        "name": "product_name",
                        "in": "path",
                        "description": "Product name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful filter products by name"
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            }
        },
        "/products/addProduct": {
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "add new product",
                "operationId": "fb94f2fc0a0e3986f25a7931fa2e8a9b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "more_period",
                                    "more_percent",
                                    "between_percent",
                                    "less_period",
                                    "less_percent",
                                    "name",
                                    "image_url",
                                    "price",
                                    "expiration_date",
                                    "category_id",
                                    "count",
                                    "description"
                                ],
                                "properties": {
                                    "more_period": {
                                        "type": "integer",
                                        "example": "30"
                                    },
                                    "more_percent": {
                                        "type": "integer",
                                        "example": "30"
                                    },
                                    "between_percent": {
                                        "type": "integer",
                                        "example": "50"
                                    },
                                    "less_period": {
                                        "type": "integer",
                                        "example": "15"
                                    },
                                    "less_percent": {
                                        "type": "integer",
                                        "example": "70"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "oil"
                                    },
                                    "image_url": {
                                        "type": "string",
                                        "format": "uri",
                                        "example": "https://example.com/image.jpg"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": "10000"
                                    },
                                    "expiration_date": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2024-10-16 08:00:00"
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "count": {
                                        "type": "integer",
                                        "example": "10"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "good oil"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful added",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "product added seccessfully"
                                        },
                                        "product": {
                                            "type": "string",
                                            "example": "[]"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
        "/products/deleteProduct/{product_id}": {
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "delete product",
                "operationId": "16e5968a531592cf053ada2ce798fbe6",
                "parameters": [
                    {
                        "name": "product_id",
                        "in": "path",
                        "description": "Product number",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "product deleted seccessfully"
                                        },
                                        "product": {
                                            "type": "string",
                                            "example": "[]"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "scheme": "bearer",
                "in": "header",
                "name": "Authorization",
                "description": "Enter your Bearer token in the format `YOUR_TOKEN`. For example: `eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...`"
            }
        }
    }
}